feat(v4): EventStudyResults consumers + dCDH aggregate shim (2(b) PR-1, M-026; M-092/M-093 amended)
v4 program section 9 item 5, PR-1 of the 2(b) wave: the unified
event-study container becomes CONSUMABLE (the TODO consumability row's
acceptance gates), and dCDH - the one row whose fit-time aggregate=
never computed anything - ships its post-fit aggregate() view.

Container consumers (M-092/M-093 pre-cut half):
- compute_honest_did, compute_pretrends_power and plot_event_study
  accept EventStudyResults from
  CallawaySantAnnaResults.aggregate('event_study');
  plot_honest_event_study renders HonestDiDResults computed from a
  container. Route parity is the pinned gate: HonestDiD bounds equal
  across routes at 1e-14; pretrends extraction bit-exact (its MVN-CDF
  power carries scipy's internal jitter either way).
- Admission in HonestDiD/PreTrendsPower is SOURCE-SCOPED to CS-produced
  containers; dCDH l1 containers rejected BY DESIGN (placebo semantics
  need the native branch); plotters take no source guard.
- EventStudyResults gains four provenance fields declared last:
  base_period, anticipation, df_survey (scalar inference df with the
  replicate-undefined 0.0 sentinel the per-row df column cannot
  encode), and reference_event_times (distinct per-cohort
  positional-base event times, computed at fit under universal base and
  recomputed per-aggregation over retained cohorts under balance_e).
  Threaded by all builders incl. the requested-but-empty path;
  to_dict() JSON-safe.

Correctness hardening that landed with the wave (both routes each):
- Common-reference guard: universal-base fits whose cohorts carry more
  than one positional-base event time (gapped grids - including the
  overlap layout where a cohort's base coincides with another cohort's
  estimated horizon and NO reference-only row marks it) fail closed in
  HonestDiD and PreTrendsPower; provenance-less universal inputs derive
  the bases from materialized reference cells (pre-3.9 pickles) or warn
  fail-safe (hand-built containers). Multi-reference containers also
  fail closed in HonestDiD (deliberate deviation from the fit-time
  first-marker split).
- Roth-anchored linear violations: pretrends relative times are
  measured from the omitted reference (|t - t_ref|) on the
  CS-universal, SunAbraham and container routes, matching the MPD
  branch's existing anchor; universal-base CS/SA linear power/MDV
  numbers change to the Roth-correct values (REGISTRY linear-convention
  note records the correction). Varying-base input keeps raw labels
  under a new interpretation warning scoped to the linear benchmark.
- Anticipation boundary unified: HonestDiD's no-reference split, the
  plotters' pre-shading, and pretrends' cutoff all treat [e=-k, -1] as
  anticipated treatment effects (e < -k is the clean pre set).
- Zero-SE rows drop on every HonestDiD adapter (MPD/dCDH joined CS/
  container); MPD extraction fails closed when a dropped horizon breaks
  the positional grid and returns the estimated period labels beta_hat
  was built from; explicit empty pre_periods subsets are rejected.
- Plotting: multi-reference containers render every anchor hollow via
  row-aligned reference marks (explicit renormalization around a
  non-anchor fails closed; REGISTRY Event Study Plotting edge case);
  interval bars draw endpoint-based so stored percentile/bootstrap
  intervals that exclude the point estimate render exactly; container
  stored-interval levels are honesty-checked against the requested
  alpha; the honest plotter restricts to HonestDiD-retained rows with
  the reference kept as a hollow normalization anchor.

dCDH shim (M-026, phase 5):
- fit(aggregate=) is deprecated: supplying it warns; any non-None value
  raises pointing at results.aggregate() (it never computed anything -
  "reserved for Phase 3").
- ChaisemartinDHaultfoeuilleResults.aggregate() ships as a pure VIEW:
  'event_study' returns the unified container (Phase-1 L_max=None fits
  the 2-row l=1 view); 'simple' a one-row AggregationResult relaying
  the overall estimand bit-exactly with estimand-aware target/n/n_kind
  (DID_M/switcher_cells, DID_1/groups, delta/no-scalar-count, the
  trends-linear all-NaN first-difference row) and df resolved from the
  ACTUAL inference path (the L_max>=2 delta keeps analytical survey df
  under bootstrap per the REGISTRY delta-SE note). Because nothing is
  recomputed, bootstrap fits are permitted - unlike CallawaySantAnna's
  kit-based aggregate(), which stays fail-closed.
- AggregationMixin routing hooks are ClassVar (the annotated override
  had leaked into the dataclass __init__/repr/eq surface); dynamic
  regression enrolls every dataclass mixin adopter.

Ledger/docs: M-026 flips to shimmed; M-092 gains its provenance-field
amendments; M-093 records the pre-cut consumer delivery with
source-scoped admission widening per-estimator; v4-design section 5 and
REGISTRY notes updated (dCDH view Note, common-reference guard Note,
plotting edge cases, pretrends linear-convention correction);
autosummary rosters extended; TODO consumability row delivered and
removed, two follow-up methodology rows added (varying-base violation
transformation; the pre-existing native-route items resolved in-tree).

+++ diff_diff/utils.py
@@ -33,10 +33,10 @@ _OPTIMIZATION_TOL = 1e-8  # Convergence tolerance for optimization
 _NUMERICAL_EPS = 1e-10  # Small constant to prevent division by zero
 
 # Cache for critical values to avoid repeated scipy calls
-_critical_value_cache: Dict[Tuple[float, Optional[int]], float] = {}
+_critical_value_cache: Dict[Tuple[float, Optional[float]], float] = {}
 
 
-def _get_critical_value(alpha: float, df: Optional[int] = None) -> float:
+def _get_critical_value(alpha: float, df: Optional[float] = None) -> float:
     """Return cached critical value for (alpha, df) pair."""
     key = (alpha, df)
     if key not in _critical_value_cache:
